fix: conform to the nbformat specification
