+++ packages/libro-code-cell/src/code-cell-view.tsx
@@ -19,6 +19,7 @@ import {
   ViewRender,
   watch,
   Deferred,
+  ConfigurationService,
 } from '@difizen/libro-common/app';
 import { l10n } from '@difizen/libro-common/l10n'; /* eslint-disable react-hooks/exhaustive-deps */
 import type {
@@ -190,10 +191,11 @@ export class LibroCodeCellView extends LibroEditableExecutableCellView {
   constructor(
     @inject(ViewOption) options: CellViewOptions,
     @inject(CellService) cellService: CellService,
+    @inject(ConfigurationService) configurationService: ConfigurationService,
     @inject(ViewManager) viewManager: ViewManager,
     @inject(CodeEditorManager) codeEditorManager: CodeEditorManager,
   ) {
-    super(options, cellService);
+    super(options, cellService, configurationService);
     this.options = options;
     this.viewManager = viewManager;
     this.codeEditorManager = codeEditorManager;
