+++ packages/libro-core/src/libro-setting.ts
@@ -119,3 +119,14 @@ export const SpmReporter: ConfigurationNode<boolean> = {
     type: 'boolean',
   },
 };
+
+export const LargeOutputDisplay: ConfigurationNode<boolean> = {
+  id: 'libro.output.large.display',
+  description: l10n.t('是否优化长文本 output 显示'),
+  title: 'Output',
+  type: 'checkbox',
+  defaultValue: true,
+  schema: {
+    type: 'boolean',
+  },
+};
