+++ packages/libro-output/src/stream-output/index.ts
@@ -1,4 +1,3 @@
 export * from './stream-output-contribution.js';
 export * from './stream-output-model.js';
 export * from './stream-output-module.js';
-export * from './output-configuration.js';
