+++ packages/libro-output/src/stream-output/stream-output-model.tsx
@@ -1,11 +1,6 @@
 import type { JSONObject } from '@difizen/libro-common';
 import { getBundleOptions } from '@difizen/libro-common';
-import {
-  ConfigurationService,
-  inject,
-  prop,
-  transient,
-} from '@difizen/libro-common/app';
+import { inject, transient } from '@difizen/libro-common/app';
 import {
   getOrigin,
   useInject,
@@ -20,7 +15,6 @@ import type { IRenderMimeRegistry, IRendererFactory } from '@difizen/libro-rende
 import { forwardRef, useEffect, useState } from 'react';
 
 import '../index.less';
-import { LargeOutputDisplay } from './output-configuration.js';
 
 const StreamOutputModelRender = forwardRef<HTMLDivElement>(
   function StreamOutputModelRender(_props, ref) {
@@ -52,34 +46,13 @@ const StreamOutputModelRender = forwardRef<HTMLDivElement>(
 export class StreamOutputModel extends LibroOutputView implements BaseOutputView {
   @inject(RenderMimeRegistry) renderMimeRegistry: IRenderMimeRegistry;
   renderFactory?: IRendererFactory;
-  @prop()
-  isLargeOutputDisplay: boolean;
 
-  constructor(
-    @inject(ViewOption) options: IOutputOptions,
-    @inject(ConfigurationService) configurationService: ConfigurationService,
-  ) {
+  constructor(@inject(ViewOption) options: IOutputOptions) {
     super(options);
     const { data, metadata } = getBundleOptions(options.output);
     this.type = options.output.output_type;
     this.data = data as JSONObject;
     this.metadata = metadata;
-    const localIsLargeOutputDisplay = localStorage.getItem(
-      'is-libro-large-output-display',
-    );
-    configurationService
-      .get(LargeOutputDisplay)
-      .then((value: any) => {
-        this.isLargeOutputDisplay =
-          localIsLargeOutputDisplay !== null
-            ? localIsLargeOutputDisplay === 'true'
-            : (options.output['is_large_display'] as boolean) || true;
-        this.isLargeOutputDisplay = this.isLargeOutputDisplay && value;
-        return;
-      })
-      .catch(() => {
-        //
-      });
   }
 
   getRenderFactory() {
@@ -97,7 +70,6 @@ export class StreamOutputModel extends LibroOutputView implements BaseOutputView
       output_type: this.raw.output_type,
       name: this.raw.name,
       text: this.raw.text,
-      is_large_display: this.isLargeOutputDisplay,
     };
   }
 }
